+++ .eslintrc.js
@@ -1,9 +1,25 @@
 module.exports = {
   extends: "react-app",
+  plugins: ["simple-import-sort"],
   parserOptions: {
     ecmaFeatures: {
       legacyDecorators: true,
     },
   },
-  rules: {},
+  rules: {
+    "simple-import-sort/sort": [
+      "error",
+      {
+        groups: [
+          ["^\\u0000"],
+          ["^@?\\w"],
+          [
+            "^@components(/.*|$)|@images(/.*|$)|@stores(/.*|$)|@constants(/.*|$)|@helpers(/.*|$)",
+          ],
+          ["^[^.]"],
+          ["^\\."],
+        ],
+      },
+    ],
+  },
 };