+++ src/App.tsx
@@ -1,30 +1,21 @@
-import React, { useEffect } from "react";
-import { autorun } from "mobx";
 import { observer } from "mobx-react";
+import React, { useEffect } from "react";
 
-import { useStore } from "./stores/context";
+import { Content } from "@components/content";
+import { Header } from "@components/header";
+import { useStore } from "@stores/context";
 
-const App = observer(() => {
+const App: React.FC = observer(() => {
   const store = useStore();
-  const test = autorun(() => {
-    console.log("props.test9", store.previewStore.file);
-  });
 
-  useEffect(() => test);
+  useEffect(() => {
+    store.previewCanvas.setCanvas("Dribbble");
+  });
 
   return (
     <>
-      <h1
-        onClick={() =>
-          console.log("store.fileStore.file", store.previewStore.file)
-        }
-      >
-        {store.previewStore.file || "Teeest"}
-      </h1>
-      <input
-        type="text"
-        onClick={(event) => store.previewStore.setFile(event)}
-      />
+      <Header />
+      <Content />
     </>
   );
 });