+++ CHANGELOG.md
@@ -1,3 +1,88 @@
+## [1.0.10](https://github.com/Alexandrshy/tiro/compare/v1.0.9...v1.0.10) (2020-09-25)
+
+
+### Bug Fixes
+
+* added input color ([fbee2ec](https://github.com/Alexandrshy/tiro/commit/fbee2ecb5ebb1f27597a118d8357f6c531d7469a))
+* added input component ([ddb539e](https://github.com/Alexandrshy/tiro/commit/ddb539e131563b410be0958285907e6aebcd2e88))
+
+## [1.0.9](https://github.com/Alexandrshy/tiro/compare/v1.0.8...v1.0.9) (2020-09-25)
+
+
+### Bug Fixes
+
+* added useEffect for App ([334abc0](https://github.com/Alexandrshy/tiro/commit/334abc0f223ecdca808829de07b284376f983846))
+* fixed alias ([b442ed6](https://github.com/Alexandrshy/tiro/commit/b442ed642fd6d7c6805619ced43eed75970a9460))
+
+## [1.0.8](https://github.com/Alexandrshy/tiro/compare/v1.0.7...v1.0.8) (2020-09-25)
+
+
+### Bug Fixes
+
+* added selector ([bd284da](https://github.com/Alexandrshy/tiro/commit/bd284daca339b68460b21f030070a138b6374bc6))
+
+## [1.0.7](https://github.com/Alexandrshy/tiro/compare/v1.0.6...v1.0.7) (2020-09-25)
+
+
+### Bug Fixes
+
+* added store for canvas ([c5d5415](https://github.com/Alexandrshy/tiro/commit/c5d5415799466bbda2e7ae171573f18587887e75))
+* added store for color ([435e5e7](https://github.com/Alexandrshy/tiro/commit/435e5e7c3d649773811ae6a09caa09a95f164052))
+* added store for size ([71bae8b](https://github.com/Alexandrshy/tiro/commit/71bae8b081d6553317938e2beb2ea74f6fb455ba))
+* expanded the context for the app ([9bfe8eb](https://github.com/Alexandrshy/tiro/commit/9bfe8eb1dbe0131bba69d5494b682c207c00446a))
+
+## [1.0.6](https://github.com/Alexandrshy/tiro/compare/v1.0.5...v1.0.6) (2020-09-22)
+
+
+### Bug Fixes
+
+* added range slider ([2d04e28](https://github.com/Alexandrshy/tiro/commit/2d04e28588f52f25fc9f4a420613952150c7d759))
+
+## [1.0.5](https://github.com/Alexandrshy/tiro/compare/v1.0.4...v1.0.5) (2020-09-20)
+
+
+### Bug Fixes
+
+* added footer ([0001642](https://github.com/Alexandrshy/tiro/commit/0001642f5cfcb900804efcae22757d90ca36a95e))
+* small fix for footer link and styles ([ebc957d](https://github.com/Alexandrshy/tiro/commit/ebc957db7fb24f0b675676079286aa8a8971063e))
+
+## [1.0.4](https://github.com/Alexandrshy/tiro/compare/v1.0.3...v1.0.4) (2020-09-19)
+
+
+### Bug Fixes
+
+* fixed css variables ([36139eb](https://github.com/Alexandrshy/tiro/commit/36139eb207893a20afa723291c242f3089ea760a))
+* fixed index imports ([ecd4c2a](https://github.com/Alexandrshy/tiro/commit/ecd4c2a77fd22e1eab7a5907ea8658ae80627f4d))
+* fixed stores ([a8531a7](https://github.com/Alexandrshy/tiro/commit/a8531a726558d07a243514978e2930536039a378))
+* init logo component, and add hover effect for buttons ([d941f1f](https://github.com/Alexandrshy/tiro/commit/d941f1f363e49b7c16828019e8272215d540f999))
+* init new components ([002e745](https://github.com/Alexandrshy/tiro/commit/002e745ce4687950d77c448a4c2a388c89ff7158))
+
+## [1.0.3](https://github.com/Alexandrshy/tiro/compare/v1.0.2...v1.0.3) (2020-09-19)
+
+
+### Bug Fixes
+
+* added rules for alias ([f627761](https://github.com/Alexandrshy/tiro/commit/f627761e96cd23f58186b398462b39a9bc1901d5))
+* fixed import sort rules ([0d944c2](https://github.com/Alexandrshy/tiro/commit/0d944c269517a788aff9e05cc0649425cd7c5ba8))
+* fixed manifest ([e737efa](https://github.com/Alexandrshy/tiro/commit/e737efa12372c68008dcb2994a737480da84ae81))
+
+## [1.0.2](https://github.com/Alexandrshy/tiro/compare/v1.0.1...v1.0.2) (2020-09-18)
+
+
+### Bug Fixes
+
+* added sort plugin, fix eslint error, removed comment ([9425ec2](https://github.com/Alexandrshy/tiro/commit/9425ec23dae7e1a75807cd7ef0593c06fae5e487))
+
+## [1.0.1](https://github.com/Alexandrshy/tiro/compare/v1.0.0...v1.0.1) (2020-09-18)
+
+
+### Bug Fixes
+
+* added button ([f892cf0](https://github.com/Alexandrshy/tiro/commit/f892cf068613120ebac474a7e06e38c7faa3ff2f))
+* added manifest and fixed html file ([01b760a](https://github.com/Alexandrshy/tiro/commit/01b760a71f4324928bfcf660a197794df5af13f4))
+* fixed env variables ([724b294](https://github.com/Alexandrshy/tiro/commit/724b294bbb94f1113905873dfa201114827bc5d2))
+* package.json ([bf24156](https://github.com/Alexandrshy/tiro/commit/bf241568810f3ee0e9502366736af49e1eba9748))
+
 # 1.0.0 (2020-09-13)